브라우저 로그추가

+++ workspace/MvcPj/src/sm/mvc/control/SmController.java
@@ -48,6 +48,10 @@ public class SmController extends HttpServlet {
 		
 		SmService service = SmService.getInstance();
 		ArrayList<Board> list = service.listS();
+		if(list == null){
+			String msg = "sqlException";
+			goMsgAlert(msg);
+		}
 		request.setAttribute("list", list);
 
 		String view = "list.jsp";
@@ -57,54 +61,48 @@ public class SmController extends HttpServlet {
 
 	private void selectCon(HttpServletRequest request, HttpServletResponse response)
 		throws ServletException, IOException{
-		int seq = -1;
+		int seq = 0;
 		String seqStr = request.getParameter("seq");
-		if(seqStr != null){
-			seqStr = seqStr.trim();
-			if(seqStr.length() != 0){
-				try{
-					seq = Integer.parseInt(seqStr);
-				}catch(NumberFormatException ne){
-					//에러처리
-				}
-			}else{
-				//에러처리
+		seq = check(seqStr);
+		
+		if(seq != -1) {
+			SmService service = SmService.getInstance();
+			Board selectCon = service.selectConS(seq);
+			if(selectCon == null){
+				String msg = "sqlException";
+				goMsgAlert(msg);
 			}
-		}else{
-			//에러처리
+			request.setAttribute("selectCon", selectCon);
+	
+			String view = "content.jsp";
+			RequestDispatcher rd = request.getRequestDispatcher(view);
+			rd.forward(request, response);
+		
+		}else {
+			String msg = "wrongAccess";
+			goMsgAlert(msg);
 		}
-		SmService service = SmService.getInstance();
-		Board selectCon = service.selectConS(seq);
-		request.setAttribute("selectCon", selectCon);
-
-		String view = "content.jsp";
-		RequestDispatcher rd = request.getRequestDispatcher(view);
-		rd.forward(request, response);
 	}
   
 	private void delete(HttpServletRequest request, HttpServletResponse response)
 		throws ServletException, IOException{
 		int seq = -1;
 		String seqStr = request.getParameter("seq");
-		if(seqStr != null){
-			seqStr = seqStr.trim();
-			if(seqStr.length() != 0){
-				try{
-					seq = Integer.parseInt(seqStr);
-				}catch(NumberFormatException ne){
-					//에러처리
-				}
+		seq = check(seqStr);
+		if(seq != -1){
+			SmService service = SmService.getInstance();
+			int result = service.deleteS(seq);
+			if(result != -1){
+				String msg = "delete";
+				goMsgAlert(msg);
 			}else{
-				//에러처리
+				String msg = "sqlException";
+				goMsgAlert(msg);
 			}
 		}else{
-			//에러처리
+			String msg = "wrongAccess";
+			goMsgAlert(msg);
 		}
-		
-		SmService service = SmService.getInstance();
-		service.deleteS(seq);
-		//알럿?
-		response.sendRedirect("sm.do");
 	}
 	
 	private void write(HttpServletResponse response) throws ServletException, IOException{
@@ -121,37 +119,39 @@ public class SmController extends HttpServlet {
 		Board dto = new Board(-1, writer, email, subject, content, null);
 		
 		SmService service = SmService.getInstance();
-		service.insertS(dto);
-		//알럿?
-		response.sendRedirect("sm.do");
+		int result = service.insertS(dto);
+		if(result != -1){
+			String msg = "insert";
+			goMsgAlert(msg);
+		}else{
+			String msg = "sqlException";
+			goMsgAlert(msg);
+		}
 	}
 
 	private void selUpCon(HttpServletRequest request, HttpServletResponse response)
 		throws ServletException, IOException{
 		int seq = -1;
 		String seqStr = request.getParameter("seq");
-		if(seqStr != null){
-			seqStr = seqStr.trim();
-			if(seqStr.length() != 0){
-				try{
-					seq = Integer.parseInt(seqStr);
-				}catch(NumberFormatException ne){
-					//에러처리
-				}
-			}else{
-				//에러처리
-			}
-		}else{
-			//에러처리
-		}
 		
-		SmService service = SmService.getInstance();
-		Board selUpCon = service.selUpConS(seq);
-		request.setAttribute("selUpCon", selUpCon);
-		//알럿?
-		String view = "upcontent.jsp";
-		RequestDispatcher rd = request.getRequestDispatcher(view);
-		rd.forward(request, response);
+		seq = check(seqStr);
+		if(seq != -1){
+			SmService service = SmService.getInstance();
+			Board selUpCon = service.selUpConS(seq);
+			if(selUpCon == null){
+				String msg = "sqlException";
+				goMsgAlert(msg);
+			}
+			request.setAttribute("selUpCon", selUpCon);
+			
+			String view = "upcontent.jsp";
+			RequestDispatcher rd = request.getRequestDispatcher(view);
+			rd.forward(request, response);
+
+		}else{
+			String msg = "wrongAccess";
+			goMsgAlert(msg);
+		}
 	}
 
 	private void update(HttpServletRequest request, HttpServletResponse response)
@@ -161,26 +161,42 @@ public class SmController extends HttpServlet {
 		String email = request.getParameter("email");
 		String subject = request.getParameter("subject");
 		String content = request.getParameter("content");
-		//메소드 처리하자 check();
+		
 		int seq = -1;
+		seq = check(seqStr);
+		if( seq != -1){
+			Board dto = new Board(seq, writer, email, subject, content);
+			
+			SmService service = SmService.getInstance();
+			service.updateS(dto);
+			response.sendRedirect("sm.do");
+		
+		}else{
+			String msg = "wrongAccess";
+			goMsgAlert(msg);
+		}
+	}
+	
+	//수정
+	private int check(String seqStr){
 		if(seqStr != null){
 			seqStr = seqStr.trim();
 			if(seqStr.length() != 0){
 				try{
-					seq = Integer.parseInt(seqStr);
-				}catch(NumberFormatException ne){
-					//에러처리
+					int seq = Integer.parseInt(seqStr);
+					return seq;
+				}catch(NumberFormatException ne){	
 				}
 			}else{
-				//에러처리
 			}
 		}else{
-			//에러처리
 		}
-		Board dto = new Board(seq, writer, email, subject, content);
-		
-		SmService service = SmService.getInstance();
-		service.updateS(dto);
-		response.sendRedirect("sm.do");
+		return -1;
+	}
+	private void goMsgAlert(String msg){
+		request.setAttribute("msg", msg);
+		String view = "msgAlert.jsp";
+		RequestDispatcher rd = request.getRequestDispatcher(view);
+		rd.forward(request, response);
 	}
 }
